make library engine work in edge client
