+++ packages/client/src/runtime/core/engines/library/types/Library.ts
@@ -1,5 +1,6 @@
 import type { ErrorCapturingDriverAdapter } from '@prisma/driver-adapter-utils'
 
+import { EngineConfig } from '../../common/Engine'
 import type { QueryEngineConfig } from '../../common/types/QueryEngine'
 
 export type QueryEngineInstance = {
@@ -27,7 +28,7 @@ export interface QueryEngineConstructor {
 }
 
 export interface LibraryLoader {
-  loadLibrary(): Promise<Library>
+  loadLibrary(config: EngineConfig): Promise<Library>
 }
 
 // Main
